Add suspenseful data and lazy-loaded route
This commit adds a mock API layer and a few features related to the
React 18 upgrade.

React 18 enables support for Suspense in server-side rendering.
This allows us to let the server attempt to fetch data in-render and
abort rendering if it takes too long. This is not shown in this example,
but can be seen here https://github.com/reactwg/react-18/discussions/37
in the "Demo" section.

Suspense also enables rendering lazy-loaded components, which is useful
for code-splitting routes. The ability to join SSR with code-splitting
is very useful for managing bundle size and time-to-first-paint.

In this case, because both client and server are TypeScript, we are also
able to define and share types between client and server.